+++ Assets/UI/StressBar.cs
@@ -19,7 +19,7 @@ public class StressBar : MonoBehaviour
         slider.minValue = 0;
         slider.value = 0.082f;
     }
-    public void setStress(int stress)
+    public void setStress(float stress)
     {
         //Map stress value from 0-> 100 to the range of values that the stress bar is visible for.
         slider.value = ((float)stress)*((0.755f-0.082f)/(float)100);
